migrated upload logic from `UploadCsvDialog` to `UploadWizard`
